Inline SutInitializer
